working 2 and 3

+++ PWR_CALIB/PWR_CALIB.py
@@ -27,10 +27,6 @@ from DRV.DRV_BK_PREC import *
 from DRV.DRV_EA import *
 
 #######################              ENUMS               #######################
-config_calib = {'version':      {'hw': None, 'sw': None, 'dev': None, 's_n': None},
-                'calib_data':   {'volt_hs': {'factor': None, 'offset': None, 'date': None},
-                                 'volt_ls': {'factor': None, 'offset': None, 'date': None},
-                                 'curr':    {'factor': None, 'offset': None, 'date': None}}}
 class _DEFAULTS():
     ERROR_RANGE = 100 #mV
 
@@ -69,11 +65,11 @@ class PWR_CALIB_c():
         multimeter_port (str): Serial port of the multimeter.
         epc_number (str): Serial number of the EPC.
     '''
-    def __init__(self, source_port: str, multimeter_port: str, epc_number: str) -> None:
+    def __init__(self, source_port: str, multimeter_port: str, epc_sn: str) -> None:
         self._source: DRV_EA_PS_c   = DRV_EA_PS_c(model = 'EA-PS 2384-05 B' , serial_port = source_port)
         self._meter: DRV_BK_Meter_c = DRV_BK_Meter_c(multimeter_port)
-        self._epc: DRV_EPC          = epc_number
-        self._epc_ser_number = epc_number
+        self._epc: DRV_EPC          = 0 #TODO: QUE VA??
+        self._epc_sn: int = epc_sn
 
     def SetupCalibStation(self, serial_number: str):
         ''' Sets up the calibration serial_number.
@@ -84,7 +80,7 @@ class PWR_CALIB_c():
         Raises:
             - #TODO: ¿¿¿¿????
         '''
-        self._epc_ser_number = serial_number
+        self._epc_sn = serial_number
         pass
     
     def CalibVoltHS(self) -> None:
@@ -96,8 +92,8 @@ class PWR_CALIB_c():
         Raises:
             - #TODO: ¿¿¿¿????
         '''
-        calib: pd.DataFrame = self._calibVolts(v_min = _Params_e.VOLT_HS_MIN.value, step = _Params_e.VOLT_HS_STEP.value, v_max = _Params_e.VOLT_HS_MAX.value)
-        #calib = BORRAR
+        #calib: pd.DataFrame = self._calibVolts(v_min = _Params_e.VOLT_HS_MIN.value, step = _Params_e.VOLT_HS_STEP.value, v_max = _Params_e.VOLT_HS_MAX.value)
+        calib = BORRAR
         line: tuple = self._parameters_line(calib)
         self._save_calib_data(type_cal = _CALIB_DATA_e.VOLT_HS.value, df = calib, factor = line[0], offset = line[1])
 
@@ -111,8 +107,8 @@ class PWR_CALIB_c():
         Raises:
             - #TODO: ¿¿¿¿????
         '''
-        calib: pd.DataFrame = self._calibVolts(v_min = _Params_e.VOLT_LS_MIN.value, step = _Params_e.VOLT_LS_STEP.value, v_max = _Params_e.VOLT_LS_MAX.value)
-        #calib = BORRAR
+        #calib: pd.DataFrame = self._calibVolts(v_min = _Params_e.VOLT_LS_MIN.value, step = _Params_e.VOLT_LS_STEP.value, v_max = _Params_e.VOLT_LS_MAX.value)
+        calib = BORRAR
         line: tuple = self._parameters_line(calib)
         self._save_calib_data(type_cal = _CALIB_DATA_e.VOLT_LS.value, df = calib, factor = line[0], offset = line[1])
 
@@ -175,7 +171,6 @@ class PWR_CALIB_c():
         Raises:
             - #TODO: ¿¿¿¿????
         '''
-        print(f'peta aqui:\n{df}\n')
         x = np.array(df['voltage multimeter']).tolist()
         y = np.array(df['voltage EPC']).tolist()
         
@@ -198,44 +193,37 @@ class PWR_CALIB_c():
         Raises:
             - #TODO: ¿¿¿¿????
         '''
-        name_path_info = f"./PWR_CALIB/CALIBRATIONS/epc_calib_info_{self._epc_ser_number}.yaml"
-        name_path_dates = f"./PWR_CALIB/CALIBRATIONS/epc_calib_{type_cal}_{self._epc_ser_number}.csv"
 
-    
-        if not os.path.exists(name_path_dates):
-            df.to_csv(name_path_dates, index=False)
+        name_path = f"CALIB_MAIN/DATAS/data_{self._epc_sn}"
+        name_path_info = f"{name_path}/epc_calib_info_{self._epc_sn}.yaml"
+        name_path_dates = f"{name_path}/epc_calib_{type_cal}_{self._epc_sn}.csv"
 
-        if not os.path.exists(name_path_info):
-            config = config_calib
-            write_dates = True
-        else:
-            with open(name_path_info, 'r') as file:
-                config = yaml.load(file, Loader=yaml.FullLoader)
-                if config['calib_data'][type_cal]['factor'] == None or config['calib_data'][type_cal]['offset'] == None or config['calib_data'][type_cal]['date'] == None:
-                    with_dates = False
-                    write_dates = True
-                else:
-                    with_dates = True
-            if with_dates:
-                while True:
-                    try:
-                        check = str(input(f"Calibration data for {type_cal} already exists. ¿Do you want to replace it? (y/n): "))
-                        if check == 'y' or check == 'Y':
-                            write_dates = True
-                            break
-                        elif check == 'n' or check == 'N':
-                            write_dates = False
-                            break
-                        else:
-                            print("Invalid option.")
-                    except:
+        with open(name_path_info, 'r') as file:
+            conf_dev = yaml.load(file, Loader=yaml.FullLoader)
+            if conf_dev['calib_data'][type_cal]['factor'] == None or conf_dev['calib_data'][type_cal]['offset'] == None or conf_dev['calib_data'][type_cal]['date'] == None:
+                with_dates = False
+                write_dates = True
+            else:
+                with_dates = True
+        if with_dates:
+            while True:
+                try:
+                    check = str(input(f"Calibration data for {type_cal} already exists. ¿Do you want to replace it? (y/n): "))
+                    if check == 'y' or check == 'Y':
+                        write_dates = True
+                        break
+                    elif check == 'n' or check == 'N':
+                        write_dates = False
+                        break
+                    else:
                         print("Invalid option.")
+                except:
+                    print("Invalid option.")
 
         if write_dates:
-            config['version']['s_n'] = self._epc_ser_number
-            config['calib_data'][type_cal]['factor'] = factor
-            config['calib_data'][type_cal]['offset'] = offset
-            config['calib_data'][type_cal]['date'] = time.strftime("%d/%m/%Y %H:%M:%S")
+            conf_dev['calib_data'][type_cal]['factor'] = factor
+            conf_dev['calib_data'][type_cal]['offset'] = offset
+            conf_dev['calib_data'][type_cal]['date'] = time.strftime("%d/%m/%Y %H:%M:%S")
             with open(name_path_info, 'w') as file:
-                yaml.dump(config, file)
+                yaml.dump(conf_dev, file)
                 df.to_csv(name_path_dates, index=False)
